+++ Sources/Swift/Integrations/UserFeedback/Configuration/SentryUserFeedbackFormConfiguration.swift
@@ -155,6 +155,20 @@ public final class SentryUserFeedbackFormConfiguration: NSObject {
     func fullLabelText(labelText: String, required: Bool) -> String {
         required ? labelText + " " + isRequiredLabel : labelText
     }
+    
+    /**
+     * Message shown to the user when an unexpected error happens while submitting feedback.
+     * - note: Default: `"Unexpected client error."`
+     */
+    public var unexpectedErrorText: String = "Unexpected client error."
+    
+    /**
+     * Message shown to the user when the form fails the validation.
+     * - note: Default: `"You must provide all required information before submitting. Please check the following field(s)"`
+     */
+    public var validationErrorMessage: (Bool) -> String = { multipleErrors in
+        return "You must provide all required information before submitting. Please check the following field\(multipleErrors ? "s" : ""):"
+    }
 }
 
 #endif // os(iOS) && !SENTRY_NO_UIKIT