+++ Sources/Swift/Integrations/UserFeedback/SentryUserFeedbackFormController.swift
@@ -88,11 +88,11 @@ extension SentryUserFeedbackFormController: SentryUserFeedbackFormViewModelDeleg
             
             guard case let SentryUserFeedbackFormViewModel.InputError.validationError(missing) = error else {
                 SentrySDKLog.warning("Unexpected error type.")
-                presentAlert(message: "Unexpected client error.", errorCode: 2, info: [NSLocalizedDescriptionKey: "Client error: ."])
+                presentAlert(message: config.formConfig.unexpectedErrorText, errorCode: 2, info: [NSLocalizedDescriptionKey: "Client error: ."])
                 return
             }
             
-            presentAlert(message: error.description, errorCode: 1, info: ["missing_fields": missing, NSLocalizedDescriptionKey: "The user did not complete the feedback form."])
+            presentAlert(message: error.getDescription(formConfig: config.formConfig), errorCode: 1, info: ["missing_fields": missing, NSLocalizedDescriptionKey: "The user did not complete the feedback form."])
         }
     }