+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - [HTTP Client errors](https://docs.sentry.io/platforms/apple/guides/ios/configuration/http-client-errors/) now mark sessions as errored (#6633)
 - The Package.swift has swift-tools-version set to 6.0 (#6764)
 
+### Improvements
+
+- Add options to customize UserFeedback error messages (#6790)
+
 ## 9.0.0-alpha.1
 
 ### Breaking Changes